requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
pyyaml==6.0.1
opencv-python==4.9.0.80
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
transformers==4.40.1
faiss-cpu==1.7.4
prometheus-client==0.20.0
python-json-logger==2.0.7
aiofiles==23.2.1
httpx==0.27.0
tenacity==8.2.3
redis==5.0.4
rich==13.7.1
structlog==24.1.0
scikit-image==0.22.0
blake3==0.4.1

# Auto-labeling dependencies (Zero-Copy AI features)
ultralytics==8.1.0  # YOLOv8 for object detection
librosa==0.10.1  # Audio processing for Whisper
pillow==10.2.0  # Image processing for BLIP-2
# Note: BLIP-2 and Whisper models are loaded via transformers (already included)
